6.5

+++ app/src/main/java/com/example/misha/myapplication/module/groups/GroupsPresenter.java
@@ -1,22 +1,34 @@
 package com.example.misha.myapplication.module.groups;
 
 import android.app.Activity;
+import android.content.ContentValues;
+import android.database.sqlite.SQLiteDatabase;
 import android.view.View;
+import android.widget.Toast;
 
 import androidx.fragment.app.FragmentActivity;
 
 import com.example.misha.myapplication.common.core.BaseMainPresenter;
+import com.example.misha.myapplication.data.database.AppContentProvider;
+import com.example.misha.myapplication.data.database.DatabaseHelper;
 import com.example.misha.myapplication.data.database.dao.AudienceDao;
 import com.example.misha.myapplication.data.database.dao.EducatorDao;
 import com.example.misha.myapplication.data.database.dao.LessonDao;
 import com.example.misha.myapplication.data.database.dao.SubjectDao;
 import com.example.misha.myapplication.data.database.dao.TypelessonDao;
 import com.example.misha.myapplication.entity.Groups;
-import com.example.misha.myapplication.entity.Lesson;
-import com.example.misha.myapplication.entity.Typelesson;
 import com.example.misha.myapplication.util.DataUtil;
 
 import java.util.ArrayList;
+
+import static com.example.misha.myapplication.data.database.dao.LessonDao.ID;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.ID_AUDIENCE;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.ID_EDUCATOR;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.ID_SUBJECT;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.ID_TYPE_LESSON;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.NUMBER_DAY;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.NUMBER_LESSON;
+import static com.example.misha.myapplication.data.database.dao.LessonDao.NUMBER_WEEK;
 
 public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> implements GroupsPresenterInterface {
 
@@ -24,7 +36,7 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
     private Activity context;
 
     public GroupsPresenter(FragmentActivity context) {
-        this.context=context;
+        this.context = context;
     }
 
     @Override
@@ -33,8 +45,8 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
     }
 
     public void onClickItem(String group, View v) {
-        loadAfterLoadGroups(group);
         getView().showProgressDialog();
+        loadSubjects(group);
     }
 
     @Override
@@ -68,30 +80,21 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
         );
     }
 
-    @Override
-    public void loadAfterLoadGroups(String group) {
-        getView().showProgressDialog();
-        getCompositeDisposable().add(getRepositoryManager()
-                .getSubjects(group)
-                .subscribeOn(getSchedulerProvider().io())
-                .observeOn(getSchedulerProvider().ui())
-                .subscribe(subjects -> loadSubjects(group), throwable -> {
-                    getView().hideProgressDialog();
-                    processSimpleError(throwable);
-                })
-        );
-    }
 
     public void loadSubjects(String group) {
         getCompositeDisposable().add(getRepositoryManager()
                 .getSubjects(group)
                 .subscribeOn(getSchedulerProvider().io())
                 .observeOn(getSchedulerProvider().ui())
-                .subscribe(groups -> {
-                    getView().hideProgressDialog();
-                    SubjectDao.getInstance().deleteAll();
-                    SubjectDao.getInstance().insertAll(groups);
-                    loadAudiences(group);
+                .subscribe(subjects -> {
+                    if (subjects.isEmpty()) {
+                        getView().hideProgressDialog();
+                        Toast.makeText(context, "Расписание выбранной группы еще не заполнено. Попробуйте загрузить позже", Toast.LENGTH_SHORT).show();
+                    } else {
+                        SubjectDao.getInstance().deleteAll();
+                        SubjectDao.getInstance().insertAll(subjects);
+                        loadAudiences(group);
+                    }
                 }, throwable -> {
                     getView().hideProgressDialog();
                     processSimpleError(throwable);
@@ -104,10 +107,9 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
                 .getAudiences(group)
                 .subscribeOn(getSchedulerProvider().io())
                 .observeOn(getSchedulerProvider().ui())
-                .subscribe(groups -> {
-                    getView().hideProgressDialog();
+                .subscribe(audiences -> {
                     AudienceDao.getInstance().deleteAll();
-                    AudienceDao.getInstance().insertAll(groups);
+                    AudienceDao.getInstance().insertAll(audiences);
                     loadEducators(group);
                 }, throwable -> {
                     getView().hideProgressDialog();
@@ -121,10 +123,9 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
                 .getEducators(group)
                 .subscribeOn(getSchedulerProvider().io())
                 .observeOn(getSchedulerProvider().ui())
-                .subscribe(groups -> {
-                    getView().hideProgressDialog();
+                .subscribe(educators -> {
                     EducatorDao.getInstance().deleteAll();
-                    EducatorDao.getInstance().insertAll(groups);
+                    EducatorDao.getInstance().insertAll(educators);
                     loadTypelessons(group);
                 }, throwable -> {
                     getView().hideProgressDialog();
@@ -138,10 +139,9 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
                 .getTypelessons()
                 .subscribeOn(getSchedulerProvider().io())
                 .observeOn(getSchedulerProvider().ui())
-                .subscribe(groups -> {
-                    getView().hideProgressDialog();
+                .subscribe(typelessons -> {
                     TypelessonDao.getInstance().deleteAll();
-                    TypelessonDao.getInstance().insertAll(groups);
+                    TypelessonDao.getInstance().insertAll(typelessons);
                     loadLessons(group);
                 }, throwable -> {
                     getView().hideProgressDialog();
@@ -155,10 +155,32 @@ public class GroupsPresenter extends BaseMainPresenter<GroupsFragmentView> imple
                 .getLessons(group)
                 .subscribeOn(getSchedulerProvider().io())
                 .observeOn(getSchedulerProvider().ui())
-                .subscribe(groups -> {
-                    getView().hideProgressDialog();
+                .subscribe(lessons -> {
                     LessonDao.getInstance().deleteAll();
-                    LessonDao.getInstance().insertAll(groups);
+
+                    DatabaseHelper databaseHelper = new DatabaseHelper(context);
+                    SQLiteDatabase database = databaseHelper.getReadableDatabase();
+                    ContentValues set;
+                    database.beginTransaction();
+
+                    try {
+                        for (int i = 0; i < 612; i++) {
+                            set = new ContentValues();
+                            set.put(ID, i + 1);
+                            set.put(NUMBER_WEEK, lessons.get(i).getNumber_week());
+                            set.put(NUMBER_DAY, lessons.get(i).getNumber_day());
+                            set.put(NUMBER_LESSON, lessons.get(i).getNumber_lesson());
+                            set.put(ID_SUBJECT, lessons.get(i).getId_subject());
+                            set.put(ID_AUDIENCE, lessons.get(i).getId_audience());
+                            set.put(ID_EDUCATOR, lessons.get(i).getId_educator());
+                            set.put(ID_TYPE_LESSON, lessons.get(i).getId_typelesson());
+                            database.insert(AppContentProvider.LESSONS_TABLE, null, set);
+                        }
+                        database.setTransactionSuccessful();
+                    } finally {
+                        database.endTransaction();
+                    }
+                    getView().hideProgressDialog();
                     DataUtil.hintKeyboard(context);
                     getView().openFragmentSchedule();
                 }, throwable -> {
